test/requirements.txt
pytest==8.3.4
cocotb==1.9.2
Pillow==10.4.0
